single-pool: remove minimum delegation check
this check is an artifact of a previous design. we used to
include the minimum in our pool stake total and burned tokens
when initializing so that it could not be withdrawn. later we
changed to treating it as separate from pool stake entirely.
therefore a post_pool_stake below minimum_delegation is valid
in the case where the first deposit is below 1 sol. this fix
allows such a deposit instead of returning a very frightening
error message